move into trait
