+++ src/Assets/Asset.php
@@ -794,19 +794,6 @@ class Asset implements AssetContract, Augmentable
         return $this->isImage() || $this->isSvg() || $this->isVideo();
     }
 
-    public function __get($key)
-    {
-        $value = $this->augmentedValue($key);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            $value = $value->get();
-        }
-
-        return $value;
-    }
-
     public function __call($method, $args)
     {
         $value = $this->augmentedValue($method);
