+++ src/Auth/Role.php
@@ -6,8 +6,6 @@ use Statamic\Contracts\Auth\Role as RoleContract;
 use Statamic\Contracts\Data\Augmentable;
 use Statamic\Data\HasAugmentedData;
 use Statamic\Facades;
-use Statamic\Facades\Compare;
-use Statamic\Fields\Value;
 
 abstract class Role implements RoleContract, Augmentable
 {
@@ -34,18 +32,5 @@ abstract class Role implements RoleContract, Augmentable
             'title' => $this->title(),
             'handle' => $this->handle(),
         ];
-    }
-
-    public function __get($key)
-    {
-        $value = $this->augmentedValue($key);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            $value = $value->get();
-        }
-
-        return $value;
     }
 }
