+++ src/Auth/User.php
@@ -243,19 +243,6 @@ abstract class User implements
         return $this->setPreference('locale', $locale);
     }
 
-    public function __get($key)
-    {
-        $value = $this->augmentedValue($key);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            $value = $value->get();
-        }
-
-        return $value;
-    }
-
     public function __call($method, $args)
     {
         $value = $this->augmentedValue($method);
