+++ src/Data/HasAugmentedInstance.php
@@ -3,6 +3,8 @@
 namespace Statamic\Data;
 
 use Statamic\Contracts\Data\Augmented;
+use Statamic\Facades\Compare;
+use Statamic\Fields\Value;
 
 trait HasAugmentedInstance
 {
@@ -46,5 +48,18 @@ trait HasAugmentedInstance
     public function shallowAugmentedArrayKeys()
     {
         return ['id', 'title', 'api_url'];
+    }
+
+    public function __get($key)
+    {
+        $value = $this->augmentedValue($key);
+
+        $value = $value instanceof Value ? $value->value() : $value;
+
+        if (Compare::isQueryBuilder($value)) {
+            $value = $value->get();
+        }
+
+        return $value;
     }
 }
