+++ src/Globals/Variables.php
@@ -180,19 +180,6 @@ class Variables implements Contract, Localization, Augmentable, ResolvesValuesCo
         return Facades\GlobalSet::find($this->id())->in($this->locale);
     }
 
-    public function __get($key)
-    {
-        $value = $this->augmentedValue($key);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            $value = $value->get();
-        }
-
-        return $value;
-    }
-
     public function __call($method, $args)
     {
         $value = $this->augmentedValue($method);
