+++ src/Sites/Site.php
@@ -4,8 +4,6 @@ namespace Statamic\Sites;
 
 use Statamic\Contracts\Data\Augmentable;
 use Statamic\Data\HasAugmentedData;
-use Statamic\Facades\Compare;
-use Statamic\Fields\Value;
 use Statamic\Support\Str;
 
 class Site implements Augmentable
@@ -112,18 +110,5 @@ class Site implements Augmentable
     public function __toString()
     {
         return $this->handle();
-    }
-
-    public function __get($key)
-    {
-        $value = $this->augmentedValue($key);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            $value = $value->get();
-        }
-
-        return $value;
     }
 }
